+++ src/main/java/com/flavorsujung/isthereopen/domain/entity/BarInfoReview.java
@@ -19,8 +19,8 @@ public class BarInfoReview {
     @Id
     @GeneratedValue(strategy= GenerationType.IDENTITY)
     Integer seq;
-    Integer userSeq;
-    Integer barSeq;
+    Long userSeq;
+    Long barSeq;
     Integer openStyle; //0 잘 지키는 편, 1 보통, 2 들쭉날쭉 지맘대로
     Integer cleanness;
     @Convert(converter = MoodConverter.class)
