+++ src/main/java/com/flavorsujung/isthereopen/domain/entity/CafeInfoReview.java
@@ -15,9 +15,9 @@ import javax.persistence.*;
 public class CafeInfoReview {
     @Id
     @GeneratedValue(strategy= GenerationType.IDENTITY)
-    Integer seq;
-    Integer userSeq;
-    Integer cafeSeq;
+    Long seq;
+    Long userSeq;
+    Long cafeSeq;
     Integer openStyle; //0 잘 지키는 편, 1 보통, 2 들쭉날쭉 지맘대로
     Integer waitingTime; // 0 짧은 편, 1 보통, 2 긴편
     Integer price; //0 쌈, 1 보통, 2 비쌈
