+++ src/main/java/com/flavorsujung/isthereopen/domain/entity/CafePatron.java
@@ -15,7 +15,7 @@ import javax.persistence.*;
 public class CafePatron {
     @Id
     @GeneratedValue(strategy= GenerationType.IDENTITY)
-    Integer seq;
-    Integer cafeSeq;
-    Integer userSeq;
+    Long seq;
+    Long cafeSeq;
+    Long userSeq;
 }
