+++ src/main/java/com/flavorsujung/isthereopen/domain/entity/RestaurantInfoReview.java
@@ -16,9 +16,9 @@ import javax.persistence.*;
 public class RestaurantInfoReview {
     @Id
     @GeneratedValue(strategy= GenerationType.IDENTITY)
-    Integer seq;
-    Integer userSeq;
-    Integer restaurantSeq;
+    Long seq;
+    Long userSeq;
+    Long restaurantSeq;
     Integer openStyle; //0 잘 지키는 편, 1 보통, 2 들쭉날쭉 지맘대로
     Integer waitingTime; // 0 짧은 편, 1 보통, 2 긴편
     Integer cleanness; //0 깨끗함, 1 보통, 2 더러움
