+++ src/main/java/com/flavorsujung/isthereopen/domain/req/ReqBarOpenReviewCreate.java
@@ -9,7 +9,7 @@ import java.util.Date;
 @Getter
 @Setter
 public class ReqBarOpenReviewCreate {
-    Integer userSeq;
+    Long userSeq;
     Integer openState; //0 close, 1 break time, 2 open, 3 미확인
     Date lastUpdate;
 }
