+++ src/main/java/com/flavorsujung/isthereopen/domain/req/ReqCafeOpenReviewCreate.java
@@ -8,7 +8,7 @@ import java.util.Date;
 @Getter
 @Setter
 public class ReqCafeOpenReviewCreate {
-    Integer userSeq;
+    Long userSeq;
     Integer openState; //0 close, 1 break time, 2 open, 3 미확인
     Date lastUpdate;
 }
