+++ src/main/java/com/flavorsujung/isthereopen/respository/BarRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface BarRepository extends JpaRepository<Bar, Integer> {
+public interface BarRepository extends JpaRepository<Bar, Long> {
     //List<Bar> findAllBar();
-    Bar findBarBySeq(Integer barSeq);
+    Bar findBarBySeq(Long barSeq);
 }
