+++ src/main/java/com/flavorsujung/isthereopen/respository/CafePatronRepository.java
@@ -4,5 +4,5 @@ import com.flavorsujung.isthereopen.domain.entity.Cafe;
 import com.flavorsujung.isthereopen.domain.entity.CafePatron;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CafePatronRepository extends JpaRepository<CafePatron, Integer> {
+public interface CafePatronRepository extends JpaRepository<CafePatron, Long> {
 }
