+++ src/main/java/com/flavorsujung/isthereopen/respository/RestaurantPatronRepository.java
@@ -4,5 +4,5 @@ import com.flavorsujung.isthereopen.domain.entity.Cafe;
 import com.flavorsujung.isthereopen.domain.entity.RestaurantPatron;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RestaurantPatronRepository extends JpaRepository<RestaurantPatron, Integer> {
+public interface RestaurantPatronRepository extends JpaRepository<RestaurantPatron, Long> {
 }
